Split status markdown agent lane fixtures
